Introduced the network-mux package
Which contains:
  * multiplexer
  * various bearers (queues, pipes and sockets)
  * channels

This package can be made independent of 'typed-protocols' if it provides
its own 'Channel' type.
