+++ network-mux/network-mux.cabal
@@ -1,0 +1,95 @@
+name:                network-mux
+version:             0.1.0.0
+synopsis:            Multiplexing library
+license:             MIT
+license-file:        LICENSE
+author:              Duncan Coutts, Marc Fontaine, Karl Knutsson, Marcin Szamotulski, Alexander Vieth
+maintainer:          duncan@well-typed.com, marcin.szamotulski@iohk.io, marc.fontaine@iohk.io, karl.knutsson@iohk.io, alex@well-typed.com
+copyright:           2019 IOHK
+category:            Network
+build-type:          Simple
+extra-source-files:  CHANGELOG.md
+cabal-version:       >=1.10
+
+Flag ipv6
+  Description: Enable IPv6 test cases
+  Manual: True
+  -- Default to False since travis lacks IPv6 support
+  Default: False
+
+library
+  build-depends:       base            >=4.9 && <4.13,
+                       typed-protocols >=0.1 && <0.2,
+                       io-sim-classes  >=0.1 && <0.2,
+                       contra-tracer   >=0.1 && <0.2,
+
+                       array           >=0.5 && <0.6,
+                       binary          >=0.8 && <0.9,
+                       bytestring      >=0.10 && <0.11,
+                       cborg           >=0.2.1 && <0.3,
+                       network,
+                       process         >=1.6 && <1.7,
+                       time            >=1.6 && <1.10
+
+  ghc-options:         -Wall
+                       -Wno-unticked-promoted-constructors
+                       -fno-ignore-asserts
+  hs-source-dirs:      src
+  exposed-modules:     Network.Mux
+                       Network.Mux.Channel
+                       Network.Mux.Egress
+                       Network.Mux.Ingress
+                       Network.Mux.Interface
+                       Network.Mux.Time
+                       Network.Mux.Types
+                       Network.Mux.Bearer.Pipe
+                       Network.Mux.Bearer.Queues
+                       Network.Mux.Bearer.Socket
+  default-language:    Haskell2010
+
+test-suite test-network-mux
+  type:                exitcode-stdio-1.0
+  hs-source-dirs:      test src
+  main-is:             Main.hs
+  other-modules:       Network.Mux
+                       Network.Mux.Channel
+                       Network.Mux.Egress
+                       Network.Mux.Ingress
+                       Network.Mux.Interface
+                       Network.Mux.Time
+                       Network.Mux.Types
+                       Network.Mux.Bearer.Pipe
+                       Network.Mux.Bearer.Queues
+                       Network.Mux.Bearer.Socket
+
+                       Test.Mux
+  default-language:    Haskell2010
+  build-depends:       base,
+                       typed-protocols,
+                       typed-protocols-cbor,
+                       io-sim-classes,
+                       io-sim            >=0.1 && < 0.2,
+                       contra-tracer,
+
+                       array,
+                       binary,
+                       bytestring,
+                       cborg,
+                       containers,
+                       hashable,
+                       network,
+                       process,
+                       QuickCheck,
+                       splitmix,
+                       serialise,
+                       stm,
+                       tasty,
+                       tasty-quickcheck,
+                       tasty-hunit,
+                       time
+
+  ghc-options:         -Wall
+                       -Wno-unticked-promoted-constructors
+                       -fno-ignore-asserts
+  if flag(ipv6)
+    cpp-options:       -DOUROBOROS_NETWORK_IPV6
